refactor: rename codecity/ → api/ (decouple dir from product name)
