+++ README.md
@@ -72,7 +72,7 @@ Two trees, cleanly separated: Python lives at the repo root, the frontend lives 
 git clone https://github.com/thalida/codecity.git
 cd codecity
 just setup                       # uv sync + npm install
-( cd web && npm run build )      # → codecity/static/
+( cd web && npm run build )      # → api/static/
 uv run codecity                  # smoke test (pick this repo in the source picker)
 ```
 
@@ -91,12 +91,12 @@ just dev           # Vite + Python API on auto-selected free ports, opens browse
 uv run pytest                    # pytest  (run from repo root)
 ```
 
-`pytest` includes a **drift check** (`codecity/tests/test_drift.py`) that does a fresh `npm run build` into a tempdir and fails if the result differs from the committed `codecity/static/`. That guarantees the bundled frontend on PyPI matches `web/` source. The check skips automatically when `npm` or `web/node_modules/` are missing.
+`pytest` includes a **drift check** (`api/tests/test_drift.py`) that does a fresh `npm run build` into a tempdir and fails if the result differs from the committed `api/static/`. That guarantees the bundled frontend on PyPI matches `web/` source. The check skips automatically when `npm` or `web/node_modules/` are missing.
 
 ### Layout
 
 ```text
-codecity/                # python package
+api/                     # python package
   cli.py                 # argparse + dispatcher
   scan.py                # filesystem + git walker
   server.py              # stdlib http server + /api routes
@@ -118,12 +118,12 @@ Cut a release from `main` after the drift test is green:
 ```sh
 # 1. Rebuild the frontend if web/ has changed since the last commit.
 ( cd web && npm run build )
-git add codecity/static
+git add api/static
 git commit -m "chore: rebuild frontend"   # only if anything changed
 
 # 2. Bump the version in BOTH places (they must match):
 #    pyproject.toml     →  version = "X.Y.Z"
-#    codecity/__init__.py →  __version__ = "X.Y.Z"
+#    api/__init__.py     →  __version__ = "X.Y.Z"
 git commit -am "chore: release vX.Y.Z"
 git tag vX.Y.Z
 
@@ -137,7 +137,7 @@ uv publish
 git push && git push --tags
 ```
 
-Why two version strings? `pyproject.toml` is the source of truth for `pip` / `uv` install resolution; `codecity/__init__.py.__version__` is what `codecity --version` prints at runtime. Keeping them in lockstep is a manual contract — drift here would surface as the CLI reporting a stale version after install.
+Why two version strings? `pyproject.toml` is the source of truth for `pip` / `uv` install resolution; `api/__init__.py.__version__` is what `codecity --version` prints at runtime. Keeping them in lockstep is a manual contract — drift here would surface as the CLI reporting a stale version after install.
 
 ## License
