+++ api/__main__.py
@@ -1,4 +1,4 @@
-from codecity.cli import main
+from api.cli import main
 
 if __name__ == "__main__":
     main()
