+++ web/.prettierignore
@@ -1,3 +1,3 @@
 node_modules
-../codecity/static
+../api/static
 package-lock.json
